:art: Onboarding View Model, Onboarding page type 분리
