+++ PyeonHaeng-iOS/Sources/Scenes/OnboardingScene/OnboardingView.swift
@@ -7,71 +7,6 @@
 
 import SwiftUI
 
-// MARK: - OnboardingPage
-
-/// 온보딩 페이지 정보를 관리하는 enum
-private enum OnboardingPage: Int, CaseIterable {
-  case first
-  case second
-
-  var title: String {
-    switch self {
-    case .first:
-      "하나사면 하나 더"
-    case .second:
-      "수많은 혜택을 한곳에서"
-    }
-  }
-
-  var body: String {
-    switch self {
-    case .first:
-      "다양한 편의점의 1+1, 2+1 행사 제품 정보로\n알뜰하고 합리적으로 소비할수 있어요."
-    case .second:
-      "세븐일레븐, CU, 이마트 24, GS 25, 미니스톱의\n수많은 행사정보를 ‘편행’한곳에서 만나보세요."
-    }
-  }
-
-  var image: Image {
-    switch self {
-    case .first:
-      Image("Onboarding01")
-    case .second:
-      Image("Onboarding02")
-    }
-  }
-
-  /// 온보딩 이미지의 상단 Spacer의 높이
-  var spacerHeight: CGFloat {
-    switch self {
-    case .first:
-      82
-    case .second:
-      139
-    }
-  }
-}
-
-// MARK: - OnboardingViewModel
-
-class OnboardingViewModel: ObservableObject {
-  @Published var currentPage: Int = 0
-
-  fileprivate var currentTitle: String {
-    OnboardingPage(rawValue: currentPage)?.title ?? ""
-  }
-
-  fileprivate var currentBody: String {
-    OnboardingPage(rawValue: currentPage)?.body ?? ""
-  }
-
-  fileprivate var nextButtonText: String {
-    currentPage < OnboardingPage.allCases.count - 1 ? "다음" : "편행 시작하기"
-  }
-
-  fileprivate var skipButtonText: String = "건너뛰기"
-}
-
 // MARK: - OnboardingView
 
 struct OnboardingView: View {
@@ -81,8 +16,8 @@ struct OnboardingView: View {
     NavigationStack {
       VStack {
         TabView(selection: $viewModel.currentPage) {
-          ForEach(OnboardingPage.allCases.indices, id: \.self) { index in
-            let page = OnboardingPage(rawValue: index) ?? OnboardingPage.first
+          ForEach(OnboardingPageType.allCases.indices, id: \.self) { index in
+            let page = OnboardingPageType(rawValue: index) ?? OnboardingPageType.first
             VStack {
               Spacer().frame(height: page.spacerHeight)
 
@@ -99,7 +34,7 @@ struct OnboardingView: View {
         .tabViewStyle(PageTabViewStyle(indexDisplayMode: .never))
 
         // 커스텀 페이지 컨트롤
-        CustomPageControl(currentPage: $viewModel.currentPage, pageCount: OnboardingPage.allCases.count)
+        CustomPageControl(currentPage: $viewModel.currentPage, pageCount: OnboardingPageType.allCases.count)
           .padding(.vertical)
 
         // 본문 제목
@@ -123,7 +58,7 @@ struct OnboardingView: View {
 
         Button(action: {
           withAnimation {
-            if viewModel.currentPage < OnboardingPage.allCases.count - 1 {
+            if viewModel.currentPage < OnboardingPageType.allCases.count - 1 {
               viewModel.currentPage += 1
             }
           }
